turn MarkTokenizer into TrialsManager
