+++ nsds_lab_to_nwb/components/stimulus/tokenizers/tone_tokenizer.py
@@ -18,12 +18,8 @@ class ToneTokenizer(BaseTokenizer):
                         ('frq', 'Stimulus Frequency'),
                         ('amp', 'Stimulus Amplitude')]
 
-    def _tokenize(self, stim_vals, mark_dset,
+    def _tokenize(self, stim_vals, stim_onsets,
                   *, stim_dur, bl_start, bl_end, rec_end_time):
-        """
-        """
-        stim_onsets = self.__get_stim_onsets(mark_dset)
-
         trial_list = []
 
         # Add the pre-stimulus period to baseline
@@ -55,7 +51,7 @@ class ToneTokenizer(BaseTokenizer):
 
         return trial_list
 
-    def __get_stim_onsets(self, mark_dset):
+    def _get_stim_onsets(self, mark_dset):
         mark_fs = mark_dset.rate
         mark_offset = self.stim_configs['mark_offset']
         stim_dur = self.stim_configs['duration']
